Refine ft_atoi and ft_strrchr loops
Agent-Logs-Url: https://github.com/amine-lr/libft/sessions/ce8f0d1c-e0ce-4f97-b2fc-9189100371f0

+++ ft_atoi.c
@@ -37,5 +37,5 @@ int	ft_atoi(const char *str)
 		res = (res * 10) + (str[i] - '0');
 		i++;
 	}
-	return ((int)(res * s));
+	return (res * s);
 }
